requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
distro==1.9.0
Django==5.2.1
django-cors-headers==4.7.0
django-ninja==1.4.1
fire==0.7.0
groq==0.25.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
python-decouple==3.8
PyYAML==6.0.2
rav==0.0.9
requests==2.32.3
rich==14.0.0
sniffio==1.3.1
sqlparse==0.5.3
termcolor==3.1.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
